Use bbqueue as buffer for RadioProcessor
The buffer implementation in src/cb.rs was causing problems at the end
of the buffer. When it tried to find a slice with a given length to
store some data, but the remaining space at the end of the buffer was
not large enough, a panic occured, even if enough space would have been
available at the start of the buffer.

This commit replaces the faulty buffer with bbqueue, which also
implements a ring buffer wher one can directly write to continuous
slices.