Cargo: update dependencies
Use latest versions where possible.  In same places dropped the
dot release specifier to take anything, and be a bit more clear that we
don't _really_ care.

Only tested with removing the Cargo.lock file and
rebuilding.  Can't update cortex-m as there's a dual dependency
somewhere in the tree and 0.6 and 0.7 clash?  Not enough of a rust
expert to resolve that.

Signed-off-by: Karl Palsson <karlp@etactica.com>